QuickTaskAPI/Database/requirements.txt
faker==19.13.0
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9 